refactor(headless-client): extract `ipc_service` module
