+++ rust/headless-client/src/linux.rs
@@ -72,26 +72,6 @@ pub(crate) fn check_token_permissions(path: &Path) -> Result<()> {
     Ok(())
 }
 
-/// Cross-platform entry point for systemd / Windows services
-///
-/// Linux uses the CLI args from here, Windows does not
-pub(crate) fn run_ipc_service(cli: CliCommon) -> Result<()> {
-    let _handle = crate::setup_ipc_service_logging(cli.log_dir)?;
-    if !nix::unistd::getuid().is_root() {
-        anyhow::bail!("This is the IPC service binary, it's not meant to run interactively.");
-    }
-    let rt = tokio::runtime::Runtime::new()?;
-    rt.spawn(crate::heartbeat::heartbeat());
-    if let Err(error) = rt.block_on(crate::ipc_listen()) {
-        tracing::error!(?error, "`ipc_listen` failed");
-    }
-    Ok(())
-}
-
-pub(crate) fn install_ipc_service() -> Result<()> {
-    bail!("`install_ipc_service` not implemented and not needed on Linux")
-}
-
 pub(crate) fn notify_service_controller() -> Result<()> {
     Ok(sd_notify::notify(true, &[sd_notify::NotifyState::Ready])?)
 }
